requirements.txt
einops==0.7.0
numpy==1.23.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
timm==0.9.7
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
